fix(nodes): resolve clippy lints and REUSE compliance on video branch (#86)
- compositor/mod.rs: use f64::from() for u32->f64 widening in
  fit_rect_preserving_aspect (clippy::cast_lossless, 6 sites)
- compositor/mod.rs: drop redundant stop_reason assignment in Shutdown
  arm; the initializer at declaration already carries "shutdown"
  (rustc unused_assignments — all break paths now assign since aa0347b)
- peer.rs: drop u128->u64 cast in tracing field; tracing accepts u128
  directly (clippy::cast_possible_truncation, matches pull.rs:780)
- LICENSES/Bitstream-Vera.txt: add SPDX license text referenced by
  REUSE.toml for bundled DejaVu fonts (reuse lint)

+++ crates/nodes/src/transport/moq/peer.rs
@@ -1094,7 +1094,7 @@ impl MoqPeerNode {
                             output_pin = pin_name,
                             attempt,
                             max = MAX_RESUBSCRIBE_ATTEMPTS,
-                            backoff_ms = backoff.as_millis() as u64,
+                            backoff_ms = backoff.as_millis(),
                             "Publisher track cancelled; re-subscribing after backoff"
                         );
                         // Yield to shutdown during backoff so we don't delay
